exemplos/fullstack-fastapi/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy>=2.0
pymysql>=1.1
cryptography>=44.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
pydantic>=2.0
